requirements.txt
# LangChain and related packages (for Gemini version)
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13

# Search API (for Gemini version)
tavily-python==0.3.3

# Google Gemini (for Gemini version)
google-generativeai==0.3.2

# OpenAI (for OpenAI version - includes built-in web search)
openai>=1.50.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0

# Document generation
python-docx==1.1.0

# FastAPI backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# File parsing for Data Room Agent
PyPDF2>=3.0.1
pdfplumber>=0.10.3
openpyxl>=3.1.2
python-pptx>=0.6.23
pandas>=2.1.4
xlrd>=2.0.1
